Fix bottom nav and delete unlinked pages
Two pages were not in the sidebar and are versions of information provided
elsewhere.

In general, "order" had gotten unmaintained, so the navigation links at the
bottom of each article jumped you around randomly. I've changed them to match
the sidebar order and have each sidebar section start at a multiple of ten to
make it slightly simpler to insert new pages in the middle.