Move site/* into top level of the repository.
While some of the docs repositories have a need for having the docs in a
nested folder, for example because of higher-level parts of
the project which have their own `package.json` or because of "code"
(used for JSDoc) submodules, etc., I don't see Galaxy Docs as needing
that configuration.  In fact, even in repositories which do have that
configuration, it's something which could be avoided, so I'd like to
avoid it here too.

I'm happy to replay these changes without this restructure!