Avoid explicit definition of the CircleCI default config.
Thanks to other changes, we have arrived at a default structure for a
Node.js project, where `npm install` is the default installation step,
and `npm test` is the default test to be run.  Thankfully, this allows
us to remove the CircleCI settings we had explicitly defined here (to
accommodate the `site` directory which previously existed, and fallback
to the default CircleCI settings.

A future update to this file can also remove the `deployment` step, as
Netlify will automatically take care of the deployment, however we'll keep
this configuration for the time-being, just in case the Netlify
deployment doesn't work out (though I don't see how that would happen).