Remove `hexo-renderer-jade`, as we have no `.jade` files here.
I'm not sure where we actually render Jade files, but there are none
present in this repository.  As one of its transitive dependencies is an
aging version of `uglify-js`, which GitHub suggests has a
"vulnerability" (which wouldn't affect this site anyway, since it's
static), I'm removing it from the dependency graph.