Remove incorrect search results/engine from sidebar.
While a search would be nice for this repository, currently, it's
returning results only for the Meteor Guide and Meteor Docs, which is
super unhelpful.

We'll need to configure the doc search for Galaxy in Algolia, if we want
this to work properly.  Unfortunately, I don't have any spare cycles to
work on that right now.

This was a result of my own oversight when putting this on the same
theme and configuration as Meteor Docs and Guide; I failed to notice
that the _config.yml of this unseen parameter was going to be inherited.