Have CircleCI use Node version from .nvmrc, rather than duplicating it.
Netlify uses the `.nvmrc` file in order to determine which version of Node
it should use, whereas CircleCI uses a `machine.node.version` attribute in
its YAML configuration.

Renovate bot then comes along and (kindly) updates the `.nvmrc` to the
latest version whenever it can, but doesn't do the same for CircleCI.

By using a post-checkout command on CircleCI, it's likely that we can avoid
this duplication and just maintain a single spot to update the version (plus
have more faith in our tests and deployments matching up).